requirements.txt
google-generativeai>=0.3.0
langgraph>=0.0.10
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
tenacity>=8.2.3
gradio>=4.0.0 